scripts/requirements.txt
feedparser>=6.0.0
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
